+++ lib/tdf3/index-web.ts
@@ -1,7 +1,6 @@
 import { FileClient } from './src/FileClient.js';
 import { Binary } from './src/binary.js';
 import { DecoratedReadableStream } from './src/client/DecoratedReadableStream.js';
-import { BrowserTdfStream } from './src/client/BrowserTdfSteam.js';
 import {
   type DecryptParams,
   DecryptParamsBuilder,
@@ -10,7 +9,6 @@ import {
   EncryptParamsBuilder,
 } from './src/client/builders.js';
 import { type SessionKeys, type ClientConfig, createSessionKeys } from './src/client/index.js';
-import { type AnyTdfStream, registerModuleType } from './src/client/tdf-stream.js';
 import { type DecryptResult, type EncryptResult } from './src/crypto/declarations.js';
 import { TDF, Client, Errors } from './src/index.js';
 import {
@@ -27,12 +25,9 @@ import {
   clientType,
 } from '../src/index.js';
 
-registerModuleType(BrowserTdfStream);
-
 window.TDF = TDF;
 
 export {
-  AnyTdfStream,
   AppIdAuthProvider,
   AuthProvider,
   AuthProviders,
