--- lib/tdf3/src/client/NodeTdfStream.ts
@@ -1,19 +1,0 @@
-import { createWriteStream } from 'node:fs';
-import { Writable } from 'node:stream';
-import {
-  DecoratedReadableStream,
-  type DecoratedReadableStreamSinkOptions,
-} from './DecoratedReadableStream.js';
-
-export class NodeTdfStream extends DecoratedReadableStream {
-  override async toFile(
-    filepath: string,
-    options: BufferEncoding | DecoratedReadableStreamSinkOptions = 'utf-8'
-  ): Promise<void> {
-    const encoding =
-      (typeof options === 'string' && options) || (options && options.encoding) || 'utf-8';
-    const nodeStream = createWriteStream(filepath, { encoding, flags: 'w', highWaterMark: 1 });
-    const writer = Writable.toWeb(nodeStream);
-    await this.stream.pipeTo(writer);
-  }
-}
