--- lib/tsconfig-node-esm.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "es2020",
-    "moduleResolution": "node16",
-    "outDir": "dist/node"
-  }
-}
